% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_PCA.R
\name{biplot_PCA}
\alias{biplot_PCA}
\title{Biplot and Score-Loading plot}
\usage{
biplot_PCA(df, a, b, cat, indicator)
}
\arguments{
\item{df}{The dataframe of chemical data and scores}

\item{a}{PC, in string. e.g. 'PC1'}

\item{b}{PC, in string. e.g. 'PC1'}

\item{cat}{Category column used for plot}

\item{indicator}{The dataframe including explained variance, obtained by following code indicator<- as.data.frame(summary(pca_objects)[[6]][,1:10])}
}
\value{
Figures with biplots and Loading plot
}
\description{
Biplot and Score-Loading plot
}
