% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_plot.R
\name{loading_plot}
\alias{loading_plot}
\title{Biplot and Score-Loading plot}
\usage{
loading_plot(df, PC_num, indicator)
}
\arguments{
\item{df}{The dataframe of PCs loading, containing information about wavenumber}

\item{PC_num}{selected PC, in string, e.g. 'PC1'}

\item{indicator}{The dataframe including explained variance, obtained by following code indicator<- as.data.frame(summary(MIR_pca)[[6]][,1:10])}
}
\value{
Loading plot
}
\description{
Biplot and Score-Loading plot
}
