% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_centering.R
\name{MC}
\alias{MC}
\title{Mean centering}
\usage{
MC(df)
}
\arguments{
\item{df}{Data Frame with col names and row names}
}
\value{
Data Frame with col names an row names
}
\description{
Mean centering
}
