% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{Principle Component Analysis(PCA)}
\usage{
pca(df, PCs = 10, ifplot = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{PCs}{How many PCs showed in summary and plots}

\item{ifplot}{if plot explained variance and cumulative explained variance when run function}
}
\value{
prcomp objects
}
\description{
Principle Component Analysis(PCA)
}
