% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Calculate modeling performance}
\usage{
performance(pred, actual, cat)
}
\arguments{
\item{pred}{the prediction values}

\item{actual}{the observed/measured values}

\item{cat}{the name of regression method used, which will be used as row name in output}
}
\value{
A dataframe contains all indicators
}
\description{
This function is used to show the performance of regression methods. Several performance indicators are calculated: R2,RMSEP,RPIQ,RPD
}
\examples{
-
}
