% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_test.R
\name{preprocessing_test}
\alias{preprocessing_test}
\title{preprocessing_test}
\usage{
preprocessing_test(spec, reference, target, seed)
}
\arguments{
\item{spec}{Input spectral data df/matrix}

\item{reference}{Chemical reference data}

\item{target}{target chemical property}

\item{seed}{set.seed for repeat checking}
}
\value{
Dataframe containing performance indicators
}
\description{
This function is used to test the performance of different pre-processing test
}
