% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNV.R
\name{SNV}
\alias{SNV}
\title{Title SNV}
\usage{
SNV(df)
}
\arguments{
\item{df}{Data Frame with col names and row names}
}
\value{
Data Frame with col names an row names
}
\description{
Title SNV
}
