% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_plot.R
\name{spec_plot}
\alias{spec_plot}
\title{Spectral ploting function}
\usage{
spec_plot(x, y, cat = NULL, main = NULL, cat_name = NULL, path = NULL)
}
\arguments{
\item{x}{Wavenumbers}

\item{y}{Spectral data}

\item{cat}{the category used to plot, the default value is NULL}

\item{main}{the title of plot}

\item{cat_name}{the legend name showed on plot, the default value is NULL}

\item{path}{storing path}
}
\value{
Spectral plot
}
\description{
Spectral ploting function
}
\examples{
-
}
